First version of fixed sampling of stepper motor pos.
In fixed stepper sample clock intervals, an accumulator is
updated with the fixed position increment (represented in a 16.16 fixed point
value). That is used to create the step accuracy by sampling the 0.1 bit.

Data sent over is
   * sample count: over how many samples the increment is to be sent.
   * Delta distance per sample: update of the position accumulator.

TODO: represent the fixed values in some sort of union; use number of
fractional bits needed to keep error < 1/2 step.